Phase 2c: Move cli2/ to packages/cli2/
Move cli2/ into the packages/ directory as part of the monorepo
restructuring. This makes cli2 a workspace member alongside the
previously moved packages (cli, server, cadl-frontend, etc.).

Changes:
- Move cli2/ to packages/cli2/ via git mv
- Create packages/cli2/package.json with workspace dependencies
- Update elm.json source-directories for new depth (../../src)
- Update root package.json bin/files entries
- Update .gitignore entries for new location
- Update mise task paths (PATHS.cli2, morphir CLI paths)
- Update runtime __dirname-based paths in morphir.ts, morphir-mcp.ts
- Update DockerTemplateFile path in cli.ts
- Update cross-package references in packages/cli/morphir-elm-make.js
  and packages/cli/morphir-elm-gen.js
- Update integration test imports to use packages/cli2/ paths
- Update dependencies.test.ts relative path assertions
- Fix cli2/.gitignore to allow package.json (needed for workspace)

All 864 Elm unit tests pass. All 13 cli2 unit tests pass.
Build succeeds for all components.
