--- cli2/.gitignore
@@ -1,4 +1,0 @@
-node_modules
-morphir-elm.ts
-package-lock.json
-package.json
