+++ packages/cli2/.gitignore
@@ -1,0 +1,3 @@
+node_modules
+morphir-elm.ts
+package-lock.json
