+++ packages/cli2/cli.ts
@@ -495,7 +495,7 @@ async function writeDockerfile(
   programOpts: any
 ): Promise<void> {
   // read docker template file
-  let filePath = "./cli2/DockerTemplateFile";
+  let filePath = "./packages/cli2/DockerTemplateFile";
   let fileContent = await fsReadFile(filePath, "utf-8");
 
   // replace specific characteres with the required
