+++ packages/cli2/dependencies.ts
