+++ packages/cli2/elm.json
@@ -1,7 +1,7 @@
 {
     "type": "application",
     "source-directories": [
-        "../src",
+        "../../src",
         "src"
     ],
     "elm-version": "0.19.1",
