+++ packages/cli2/morphir-generate-test-data.ts
