+++ packages/cli2/morphir-json-schema-gen.ts
